+++ src/core/Core/Contexts/Pipeline.hs
@@ -7,14 +7,14 @@ where
 
 import Core.Contexts.ParamsEncoder qualified as ParamsEncoder
 import Core.Location
+import Core.SessionError
 import Core.Structures.StatementCache qualified as StatementCache
-import Core.UsageError
 import Hipq.ResultDecoder qualified
 import Hipq.Roundtrip qualified
 import Platform.Prelude
 import Pq qualified
 
-run :: forall a. Pipeline a -> Bool -> Pq.Connection -> StatementCache.StatementCache -> IO (Either UsageError a, StatementCache.StatementCache)
+run :: forall a. Pipeline a -> Bool -> Pq.Connection -> StatementCache.StatementCache -> IO (Either SessionError a, StatementCache.StatementCache)
 run (Pipeline totalStatements run) usePreparedStatements connection cache = do
   let (roundtrip, newCache) = run 0 usePreparedStatements cache
       adaptedRoundtrip = first adaptContext roundtrip
@@ -22,7 +22,7 @@ run (Pipeline totalStatements run) usePreparedStatements connection cache = do
   case result of
     Left (Hipq.Roundtrip.ClientError _context details) -> do
       Pq.reset connection
-      pure (Left (ConnectionUsageError (maybe "Connection error" decodeUtf8Lenient details)), StatementCache.empty)
+      pure (Left (ConnectionSessionError (maybe "Connection error" decodeUtf8Lenient details)), StatementCache.empty)
     Left (Hipq.Roundtrip.ServerError recvError) ->
       pure (Left (fromRecvError recvError), newCache)
     Right a ->