+++ src/library/Hasql/Connection.hs
@@ -5,7 +5,7 @@
 module Hasql.Connection
   ( Connection,
     AcquisitionError (..),
-    UsageError (..),
+    SessionError (..),
     acquire,
     release,
     use,
@@ -14,9 +14,9 @@ module Hasql.Connection
 where
 
 import Core.Contexts.Session qualified as Session
+import Core.SessionError
 import Core.Structures.ConnectionState qualified as ConnectionState
 import Core.Structures.StatementCache qualified as StatementCache
-import Core.UsageError
 import Data.Text qualified as Text
 import Hasql.Connection.Config qualified as Config
 import Hasql.Connection.ServerVersion qualified as ServerVersion
@@ -140,7 +140,7 @@ withLibPQConnection connection action =
 -- Execute a sequence of operations with exclusive access to the connection.
 --
 -- Blocks until the connection is available when there is another session running upon the connection.
-use :: Connection -> Session.Session a -> IO (Either UsageError a)
+use :: Connection -> Session.Session a -> IO (Either SessionError a)
 use connection session =
   useConnectionState connection \connectionState -> do
     Session.run session connectionState